+++ src/main/java/com/example/lookkit/order/OrderVO.java
@@ -11,8 +11,8 @@ import java.util.List;
 @Getter
 @Setter
 public class OrderVO {
-    private long orderId;
-    private long userId;
+    private int orderId;
+    private int userId;
     private int totalAmount;
     private String orderStatus;
     private String orderComment;
